+++ packages/client/test/end-to-end/ProxyPublishSubscribe.test.ts
@@ -81,7 +81,7 @@ describe('PubSub with proxy connections', () => {
             receivedMessagesProxy.push(msg)
         })
         await wait(SUBSCRIBE_WAIT_TIME)
-        await onewayClient.addProxyConnectionCandidates(stream, [proxyNodeId1], ProxyDirection.PUBLISH)
+        await onewayClient.setProxies(stream, [proxyNodeId1], ProxyDirection.PUBLISH)
 
         expect((await onewayClient.getNode())
             .hasProxyConnection(toStreamPartID(stream.id, 0), proxyNodeId1, ProxyDirection.PUBLISH))
@@ -114,7 +114,7 @@ describe('PubSub with proxy connections', () => {
             receivedMessagesProxy2.push(msg)
         })
         await wait(SUBSCRIBE_WAIT_TIME)
-        await onewayClient.addProxyConnectionCandidates(stream, [proxyNodeId1, proxyNodeId2], ProxyDirection.PUBLISH)
+        await onewayClient.setProxies(stream, [proxyNodeId1, proxyNodeId2], ProxyDirection.PUBLISH)
 
         expect((await onewayClient.getNode())
             .hasProxyConnection(toStreamPartID(stream.id, 0), proxyNodeId1, ProxyDirection.PUBLISH))
@@ -124,7 +124,7 @@ describe('PubSub with proxy connections', () => {
             .hasProxyConnection(toStreamPartID(stream.id, 0), proxyNodeId2, ProxyDirection.PUBLISH))
             .toEqual(true)
 
-        await onewayClient.removeProxyConnectionCandidates(stream, [proxyNodeId1, proxyNodeId2])
+        await onewayClient.setProxies(stream, [], ProxyDirection.PUBLISH)
 
         expect((await onewayClient.getNode())
             .hasStreamPart(toStreamPartID(stream.id, 0)))
@@ -141,7 +141,7 @@ describe('PubSub with proxy connections', () => {
         await proxyClient1.subscribe(stream)
         await wait(SUBSCRIBE_WAIT_TIME)
 
-        await onewayClient.addProxyConnectionCandidates(stream, [proxyNodeId1], ProxyDirection.SUBSCRIBE)
+        await onewayClient.setProxies(stream, [proxyNodeId1], ProxyDirection.SUBSCRIBE)
         await onewayClient.subscribe(stream, (msg) => {
             receivedMessages.push(msg)
         })
@@ -170,7 +170,7 @@ describe('PubSub with proxy connections', () => {
         await proxyClient1.subscribe(stream)
         await proxyClient2.subscribe(stream)
         await wait(SUBSCRIBE_WAIT_TIME)
-        await onewayClient.addProxyConnectionCandidates(stream, [proxyNodeId1, proxyNodeId2], ProxyDirection.SUBSCRIBE)
+        await onewayClient.setProxies(stream, [proxyNodeId1, proxyNodeId2], ProxyDirection.SUBSCRIBE)
 
         expect((await onewayClient.getNode())
             .hasProxyConnection(toStreamPartID(stream.id, 0), proxyNodeId1, ProxyDirection.SUBSCRIBE))
@@ -181,7 +181,7 @@ describe('PubSub with proxy connections', () => {
             .toEqual(true)
 
         await onewayClient.unsubscribe(stream)
-        await onewayClient.removeProxyConnectionCandidates(stream, [proxyNodeId1, proxyNodeId2])
+        await onewayClient.setProxies(stream, [], ProxyDirection.SUBSCRIBE)
 
         expect((await onewayClient.getNode())
             .hasStreamPart(toStreamPartID(stream.id, 0)))
@@ -202,7 +202,7 @@ describe('PubSub with proxy connections', () => {
             receivedMessagesProxy2.push(msg)
         })
         await wait(SUBSCRIBE_WAIT_TIME)
-        await onewayClient.addProxyConnectionCandidates(stream, [proxyNodeId1, proxyNodeId2], ProxyDirection.PUBLISH)
+        await onewayClient.setProxies(stream, [proxyNodeId1, proxyNodeId2], ProxyDirection.PUBLISH)
 
         expect((await onewayClient.getNode())
             .hasProxyConnection(toStreamPartID(stream.id, 0), proxyNodeId1, ProxyDirection.PUBLISH))
@@ -212,7 +212,7 @@ describe('PubSub with proxy connections', () => {
             .hasProxyConnection(toStreamPartID(stream.id, 0), proxyNodeId2, ProxyDirection.PUBLISH))
             .toEqual(true)
 
-        await onewayClient.removeAllProxyConnectionCandidates(stream)
+        await onewayClient.setProxies(stream, [], ProxyDirection.PUBLISH)
 
         expect((await onewayClient.getNode())
             .hasStreamPart(toStreamPartID(stream.id, 0)))
@@ -224,35 +224,4 @@ describe('PubSub with proxy connections', () => {
 
     }, 15000)
 
-    it('Set number of target connections', async () => {
-        const receivedMessagesProxy1: any[] = []
-        const receivedMessagesProxy2: any[] = []
-        await proxyClient1.subscribe(stream, (msg) => {
-            receivedMessagesProxy1.push(msg)
-        })
-        await proxyClient2.subscribe(stream, (msg) => {
-            receivedMessagesProxy2.push(msg)
-        })
-        await wait(SUBSCRIBE_WAIT_TIME)
-        await onewayClient.addProxyConnectionCandidates(stream, [proxyNodeId1, proxyNodeId2], ProxyDirection.PUBLISH, 1)
-
-        const hasFirst = (await onewayClient.getNode())
-            .hasProxyConnection(toStreamPartID(stream.id, 0), proxyNodeId1, ProxyDirection.PUBLISH)
-        const hasSecond = (await onewayClient.getNode())
-            .hasProxyConnection(toStreamPartID(stream.id, 0), proxyNodeId2, ProxyDirection.PUBLISH)
-
-        expect((hasFirst && hasSecond))
-            .toEqual(false)
-
-        await onewayClient.setProxyConnectionTargetCount(stream, 2)
-
-        expect((await onewayClient.getNode())
-            .hasProxyConnection(toStreamPartID(stream.id, 0), proxyNodeId1, ProxyDirection.PUBLISH))
-            .toEqual(true)
-
-        expect((await onewayClient.getNode())
-            .hasProxyConnection(toStreamPartID(stream.id, 0), proxyNodeId2, ProxyDirection.PUBLISH))
-            .toEqual(true)
-
-    }, 15000)
 })