+++ packages/network/test/integration/proxy-connections.test.ts
@@ -74,15 +74,15 @@ describe('Proxy connection tests', () => {
     })
 
     it('publisher node can form proxy connections', async () => {
-        await onewayNode.addProxyConnectionCandidates(defaultStreamPartId, ['contact-node', 'contact-node-2'], ProxyDirection.PUBLISH, 'publisher')
+        await onewayNode.setProxies(defaultStreamPartId, ['contact-node', 'contact-node-2'], ProxyDirection.PUBLISH, 'publisher')
         // @ts-expect-error private
         expect(onewayNode.streamPartManager.getOutboundNodesForStreamPart(defaultStreamPartId)).toContainValues(['contact-node', 'contact-node-2'])
         // @ts-expect-error private
         expect(onewayNode.streamPartManager.getNeighborsForStreamPart(defaultStreamPartId)).toBeArrayOfSize(0)
     })
-
+    
     it('subscriber node can form proxy connections', async () => {
-        await onewayNode.addProxyConnectionCandidates(defaultStreamPartId, ['contact-node', 'contact-node-2'], ProxyDirection.SUBSCRIBE, 'subscriber')
+        await onewayNode.setProxies(defaultStreamPartId, ['contact-node', 'contact-node-2'], ProxyDirection.SUBSCRIBE, 'subscriber')
         // @ts-expect-error private
         expect(onewayNode.streamPartManager.getInboundNodesForStreamPart(defaultStreamPartId)).toContainValues(['contact-node', 'contact-node-2'])
         // @ts-expect-error private
@@ -90,7 +90,7 @@ describe('Proxy connection tests', () => {
     })
 
     it('publisher node can close proxy connections', async () => {
-        await onewayNode.addProxyConnectionCandidates(defaultStreamPartId, ['contact-node', 'contact-node-2'], ProxyDirection.PUBLISH, 'publisher')
+        await onewayNode.setProxies(defaultStreamPartId, ['contact-node', 'contact-node-2'], ProxyDirection.PUBLISH, 'publisher')
         // @ts-expect-error private
         expect(onewayNode.streamPartManager.getOutboundNodesForStreamPart(defaultStreamPartId)).toContainValues(['contact-node', 'contact-node-2'])
         // @ts-expect-error private
@@ -98,7 +98,7 @@ describe('Proxy connection tests', () => {
 
         await Promise.all([
             waitForEvent(contactNode, NodeEvent.ONE_WAY_CONNECTION_CLOSED),
-            onewayNode.removeProxyConnectionCandidates(defaultStreamPartId, ['contact-node']),
+            onewayNode.setProxies(defaultStreamPartId, ['contact-node-2'], ProxyDirection.PUBLISH, 'publisher'),
         ])
 
         // @ts-expect-error private
@@ -107,7 +107,7 @@ describe('Proxy connection tests', () => {
         expect(onewayNode.streamPartManager.hasOutOnlyConnection(defaultStreamPartId, 'contact-node-2')).toBeTrue()
         await Promise.all([
             waitForEvent(contactNode2, NodeEvent.ONE_WAY_CONNECTION_CLOSED),
-            onewayNode.removeProxyConnectionCandidates(defaultStreamPartId, ['contact-node-2']),
+            onewayNode.setProxies(defaultStreamPartId, ['contact-node-1'], ProxyDirection.PUBLISH, 'publisher'),
         ])
 
         // @ts-expect-error private
@@ -119,8 +119,7 @@ describe('Proxy connection tests', () => {
     })
 
     it('subscriber node can close proxy connections', async () => {
-        await onewayNode.addProxyConnectionCandidates(defaultStreamPartId, ['contact-node-2'], ProxyDirection.SUBSCRIBE, 'subscriber')
-        await onewayNode.addProxyConnectionCandidates(defaultStreamPartId, ['contact-node'], ProxyDirection.SUBSCRIBE, 'subscriber', 2)
+        await onewayNode.setProxies(defaultStreamPartId, ['contact-node', 'contact-node-2'], ProxyDirection.SUBSCRIBE, 'subscriber', 2)
         // @ts-expect-error private
         expect(onewayNode.streamPartManager.getInboundNodesForStreamPart(defaultStreamPartId)).toContainValues(['contact-node', 'contact-node-2'])
         // @ts-expect-error private
@@ -128,7 +127,7 @@ describe('Proxy connection tests', () => {
 
         await Promise.all([
             waitForEvent(contactNode, NodeEvent.ONE_WAY_CONNECTION_CLOSED),
-            onewayNode.removeProxyConnectionCandidates(defaultStreamPartId, ['contact-node']),
+            onewayNode.setProxies(defaultStreamPartId, ['contact-node-2'], ProxyDirection.SUBSCRIBE, 'subscriber'),
         ])
 
         // @ts-expect-error private
@@ -137,7 +136,7 @@ describe('Proxy connection tests', () => {
         expect(onewayNode.streamPartManager.hasInOnlyConnection(defaultStreamPartId, 'contact-node-2')).toBeTrue()
         await Promise.all([
             waitForEvent(contactNode2, NodeEvent.ONE_WAY_CONNECTION_CLOSED),
-            onewayNode.removeProxyConnectionCandidates(defaultStreamPartId, ['contact-node-2']),
+            onewayNode.setProxies(defaultStreamPartId, [], ProxyDirection.SUBSCRIBE, 'subscriber'),
         ])
 
         // @ts-expect-error private
@@ -146,13 +145,6 @@ describe('Proxy connection tests', () => {
         expect(contactNode.streamPartManager.getOutboundNodesForStreamPart(defaultStreamPartId)).toContainValues(['contact-node-2'])
         // @ts-expect-error private
         expect(contactNode.streamPartManager.hasOutOnlyConnection(defaultStreamPartId, 'publisher')).toBeFalse()
-    })
-
-    it('closing one-way connections for incorrect direction rejects', async () => {
-        await onewayNode.addProxyConnectionCandidates(defaultStreamPartId, ['contact-node'], ProxyDirection.PUBLISH, 'publisher')
-        await expect(onewayNode.removeProxyConnectionCandidates(defaultStreamPartId, ['non-contact-node']))
-            .rejects
-            .toMatch(`proxy candidate`)
     })
 
     it('publisher cannot connect to non-contact node', async () => {
@@ -169,7 +161,7 @@ describe('Proxy connection tests', () => {
             nonContactNode.subscribe(defaultStreamPartId)
         ])
 
-        await expect(onewayNode.addProxyConnectionCandidates(defaultStreamPartId, ['non-contact-node'], ProxyDirection.PUBLISH, 'publisher'))
+        await expect(onewayNode.setProxies(defaultStreamPartId, ['non-contact-node'], ProxyDirection.PUBLISH, 'publisher'))
             .rejects
             .toMatchObject(
                 new Error('Could not open any initial ProxyConnections: ' +
@@ -186,17 +178,17 @@ describe('Proxy connection tests', () => {
         await Promise.all([
             // @ts-expect-error private
             waitForEvent(onewayNode.proxyStreamConnectionClient, ProxyEvent.PROXY_CONNECTION_ACCEPTED),
-            onewayNode.addProxyConnectionCandidates(defaultStreamPartId, ['contact-node'], ProxyDirection.PUBLISH, 'publisher'),
-            onewayNode.addProxyConnectionCandidates(defaultStreamPartId, ['contact-node'], ProxyDirection.PUBLISH, 'publisher'),
-            onewayNode.addProxyConnectionCandidates(defaultStreamPartId, ['contact-node'], ProxyDirection.PUBLISH, 'publisher'),
-            onewayNode.addProxyConnectionCandidates(defaultStreamPartId, ['contact-node'], ProxyDirection.PUBLISH, 'publisher'),
+            onewayNode.setProxies(defaultStreamPartId, ['contact-node'], ProxyDirection.PUBLISH, 'publisher'),
+            onewayNode.setProxies(defaultStreamPartId, ['contact-node'], ProxyDirection.PUBLISH, 'publisher'),
+            onewayNode.setProxies(defaultStreamPartId, ['contact-node'], ProxyDirection.PUBLISH, 'publisher'),
+            onewayNode.setProxies(defaultStreamPartId, ['contact-node'], ProxyDirection.PUBLISH, 'publisher'),
         ])
         // @ts-expect-error private
         expect(onewayNode.streamPartManager.getOutboundNodesForStreamPart(defaultStreamPartId)).toContainValue('contact-node')
     })
 
     it('Published data is received using proxy publish stream connections', async () => {
-        await onewayNode.addProxyConnectionCandidates(defaultStreamPartId, ['contact-node'], ProxyDirection.PUBLISH, 'publisher')
+        await onewayNode.setProxies(defaultStreamPartId, ['contact-node'], ProxyDirection.PUBLISH, 'publisher')
         await Promise.all([
             waitForEvent(contactNode, NodeEvent.MESSAGE_RECEIVED),
             waitForEvent(contactNode2, NodeEvent.MESSAGE_RECEIVED),
@@ -211,7 +203,7 @@ describe('Proxy connection tests', () => {
     })
 
     it('proxied subscribers receive data', async () => {
-        await onewayNode.addProxyConnectionCandidates(defaultStreamPartId, ['contact-node'], ProxyDirection.SUBSCRIBE, 'subscriber')
+        await onewayNode.setProxies(defaultStreamPartId, ['contact-node'], ProxyDirection.SUBSCRIBE, 'subscriber')
         await Promise.all([
             waitForEvent(onewayNode, NodeEvent.MESSAGE_RECEIVED),
             contactNode.publish(new StreamMessage({
@@ -226,7 +218,7 @@ describe('Proxy connection tests', () => {
     })
 
     it('proxied subscribers cannot publish data', async () => {
-        await onewayNode.addProxyConnectionCandidates(defaultStreamPartId, ['contact-node'], ProxyDirection.SUBSCRIBE, 'subscriber')
+        await onewayNode.setProxies(defaultStreamPartId, ['contact-node'], ProxyDirection.SUBSCRIBE, 'subscriber')
         expect(() => onewayNode.publish(new StreamMessage({
             messageId: new MessageID(toStreamID('stream-0'), 0, 120, 0, PUBLISHER_ID, 'session'),
             content: {
@@ -237,7 +229,7 @@ describe('Proxy connection tests', () => {
     })
 
     it('Cannot open a proxy publish stream connection to non-existing node (not connected to the streams tracker)', async () => {
-        await expect(onewayNode.addProxyConnectionCandidates(defaultStreamPartId, ['non-contact-node'], ProxyDirection.PUBLISH, 'publisher'))
+        await expect(onewayNode.setProxies(defaultStreamPartId, ['non-contact-node'], ProxyDirection.PUBLISH, 'publisher'))
             .rejects
             .toMatchObject(
                 new Error('Could not open any initial ProxyConnections: ' +
@@ -250,7 +242,7 @@ describe('Proxy connection tests', () => {
     })
 
     it('Cannot open a proxy subscribe stream connection to a node without an existing subscription to the given stream', async () => {
-        await expect(onewayNode.addProxyConnectionCandidates(defaultStreamPartId, ['non-contact-node'], ProxyDirection.SUBSCRIBE, 'subscriber'))
+        await expect(onewayNode.setProxies(defaultStreamPartId, ['non-contact-node'], ProxyDirection.SUBSCRIBE, 'subscriber'))
             .rejects
             .toMatchObject(
                 new Error('Could not open any initial ProxyConnections: ' +
@@ -263,18 +255,7 @@ describe('Proxy connection tests', () => {
     })
 
     it('if caught, failed publish only connections do not clean out existing connections', async () => {
-        await onewayNode.addProxyConnectionCandidates(defaultStreamPartId, ['contact-node'], ProxyDirection.PUBLISH, 'publisher', 3)
-        await onewayNode.addProxyConnectionCandidates(defaultStreamPartId, ['contact-node-2'], ProxyDirection.PUBLISH, 'publisher')
-        try {
-            await onewayNode.addProxyConnectionCandidates(
-                StreamPartIDUtils.parse('stream-5#0'),
-                ['non-existing-node'],
-                ProxyDirection.PUBLISH,
-                'publisher'
-            )
-        } catch (err) {
-            // no-op
-        }
+        await onewayNode.setProxies(defaultStreamPartId, ['contact-node', 'contact-node-2', 'non-existing-node'], ProxyDirection.PUBLISH, 'publisher')
 
         // @ts-expect-error private
         expect(onewayNode.streamPartManager.getOutboundNodesForStreamPart(defaultStreamPartId)).toContainValues(['contact-node', 'contact-node-2'])
@@ -287,27 +268,17 @@ describe('Proxy connection tests', () => {
             waitForEvent(contactNode, NodeEvent.NODE_UNSUBSCRIBED),
             contactNode2.unsubscribe(defaultStreamPartId)
         ])
-        await onewayNode.addProxyConnectionCandidates(defaultStreamPartId, ['contact-node'], ProxyDirection.PUBLISH, 'publisher')
+        await onewayNode.setProxies(defaultStreamPartId, ['contact-node'], ProxyDirection.PUBLISH, 'publisher')
         await Promise.all([
             waitForEvent(contactNode, NodeEvent.ONE_WAY_CONNECTION_CLOSED),
-            onewayNode.removeProxyConnectionCandidates(defaultStreamPartId, ['contact-node']),
+            onewayNode.setProxies(defaultStreamPartId, [], ProxyDirection.PUBLISH, 'publisher'),
         ])
         // @ts-expect-error private
         expect(contactNode.streamPartManager.isSetUp(defaultStreamPartId)).toBeTrue()
     })
 
-    it('will not connect above set target limit, one by one', async () => {
-        await onewayNode.addProxyConnectionCandidates(defaultStreamPartId, ['contact-node'], ProxyDirection.SUBSCRIBE, 'subscriber')
-        await onewayNode.addProxyConnectionCandidates(defaultStreamPartId, ['contact-node-2'], ProxyDirection.SUBSCRIBE, 'subscriber')
-
-        // @ts-expect-error private
-        expect(onewayNode.streamPartManager.getInboundNodesForStreamPart(defaultStreamPartId)).toContainValues(['contact-node'])
-        // @ts-expect-error private
-        expect(onewayNode.streamPartManager.getNeighborsForStreamPart(defaultStreamPartId)).toBeArrayOfSize(0)
-    })
-
-    it('will not connect above set target limit, simultaneous', async () => {
-        await onewayNode.addProxyConnectionCandidates(
+    it('will not connect above set target limit', async () => {
+        await onewayNode.setProxies(
             defaultStreamPartId,
             ['contact-node', 'contact-node-2'],
             ProxyDirection.SUBSCRIBE,
@@ -322,7 +293,7 @@ describe('Proxy connection tests', () => {
     })
 
     it('will connect up to a newly set target connection count', async () => {
-        await onewayNode.addProxyConnectionCandidates(
+        await onewayNode.setProxies(
             defaultStreamPartId,
             ['contact-node', 'contact-node-2'],
             ProxyDirection.SUBSCRIBE,
@@ -335,7 +306,12 @@ describe('Proxy connection tests', () => {
         // @ts-expect-error private
         expect(onewayNode.streamPartManager.getNeighborsForStreamPart(defaultStreamPartId)).toBeArrayOfSize(0)
 
-        await onewayNode.setNumberOfTargetProxyConnections(defaultStreamPartId, 2)
+        await onewayNode.setProxies(defaultStreamPartId,
+            ['contact-node', 'contact-node-2'],
+            ProxyDirection.SUBSCRIBE,
+            'subscriber',
+            2
+        )
 
         // @ts-expect-error private
         expect(onewayNode.streamPartManager.getInboundNodesForStreamPart(defaultStreamPartId)).toBeArrayOfSize(2)
@@ -344,7 +320,7 @@ describe('Proxy connection tests', () => {
     })
 
     it('will disconnect down to a newly set target connection count', async () => {
-        await onewayNode.addProxyConnectionCandidates(
+        await onewayNode.setProxies(
             defaultStreamPartId,
             ['contact-node', 'contact-node-2'],
             ProxyDirection.SUBSCRIBE,
@@ -357,8 +333,13 @@ describe('Proxy connection tests', () => {
         // @ts-expect-error private
         expect(onewayNode.streamPartManager.getNeighborsForStreamPart(defaultStreamPartId)).toBeArrayOfSize(0)
 
-        await onewayNode.setNumberOfTargetProxyConnections(defaultStreamPartId, 1)
-
+        await onewayNode.setProxies(
+            defaultStreamPartId,
+            ['contact-node', 'contact-node-2'],
+            ProxyDirection.SUBSCRIBE,
+            'subscriber',
+            1
+        )
         // @ts-expect-error private
         expect(onewayNode.streamPartManager.getInboundNodesForStreamPart(defaultStreamPartId)).toBeArrayOfSize(1)
         // @ts-expect-error private
@@ -366,7 +347,7 @@ describe('Proxy connection tests', () => {
     })
 
     it('will reconnect after lost connectivity', async () => {
-        await onewayNode.addProxyConnectionCandidates(defaultStreamPartId, ['contact-node'], ProxyDirection.PUBLISH, 'publisher')
+        await onewayNode.setProxies(defaultStreamPartId, ['contact-node'], ProxyDirection.PUBLISH, 'publisher')
 
         await Promise.all([
             waitForEvent(contactNode, NodeEvent.NODE_CONNECTED, 20000),
@@ -386,25 +367,4 @@ describe('Proxy connection tests', () => {
             }))
         ])
     }, 20100)
-
-    it('stop proxy connections on stream', async () => {
-        await onewayNode.addProxyConnectionCandidates(
-            defaultStreamPartId,
-            ['contact-node', 'contact-node-2'],
-            ProxyDirection.SUBSCRIBE,
-            'subscriber',
-            2
-        )
-
-        // @ts-expect-error private
-        expect(onewayNode.streamPartManager.getInboundNodesForStreamPart(defaultStreamPartId)).toBeArrayOfSize(2)
-        // @ts-expect-error private
-        expect(onewayNode.streamPartManager.getNeighborsForStreamPart(defaultStreamPartId)).toBeArrayOfSize(0)
-
-        await onewayNode.removeAllProxyConnectionCandidates(defaultStreamPartId)
-
-        // @ts-expect-error private
-        expect(onewayNode.streamPartManager.isSetUp(defaultStreamPartId)).toEqual(false)
-
-    })
 })